SaltedOutpointHasher uses rapidhash
SipHashUint256Extra is rather slow. For the purpose of generating a hash
from a COutPoint and some seed that is only used inside a hashmap, it is
sufficient to use a non-cryptographic hash. rapidhash [1] is a well tested
and very fast hash function. This implementation strips down this hash
function, originally implemented for a memory buffer, to be used with
the COutPoint + 2*64bit seed as the input.

[1] https://github.com/Nicoshev/rapidhash