optimization: Bulk serialization writes in `WriteBlockUndo` and `WriteBlock`
Added `AutoFile::write_buffer` for batching obfuscation operations, so instead of copying the data and doing the xor in a 4096 byte array, we're doing it directly on the input.

Similarly to the serialization reads, buffered writes will enable batched xor calculations - especially since currently we need to copy the write input's std::span to do the obfuscation on it, batching enables doing the xor on the internal buffer instead.

> cmake -B build -DBUILD_BENCH=ON -DCMAKE_BUILD_TYPE=Release -DCMAKE_C_COMPILER=clang -DCMAKE_CXX_COMPILER=clang++ && cmake --build build -j$(nproc) && build/bin/bench_bitcoin -filter='WriteBlockBench' -min-time=10000
> C++ compiler .......................... Clang 19.1.7

Before:

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|        5,149,564.31 |              194.19 |    0.8% |     10.95 | `WriteBlockBench`

After:

|               ns/op |                op/s |    err% |     total | benchmark
|--------------------:|--------------------:|--------:|----------:|:----------
|        1,708,433.73 |              585.33 |    1.4% |     11.00 | `WriteBlockBench`

> C++ compiler .......................... GNU 13.3.0

Before:

TODO

After:

TODO

Co-authored-by: Cory Fields <cory-nospam-@coryfields.com>
Co-authored-by: Ryan Ofsky <ryan@ofsky.org>