Prepare dbwrapper for obfuscation key change
Since `CDBWrapper::Read` will still work with vectors, we won't be able to use the final key to read into it directly.
This commit cleans up this part of the code, obviating that writing `obfuscate_key` is needed since following methods will actually use it implicitly, simplifying the `if (!key_exists` condition to extract the negation into the name of the boolean and inline the single-use `CreateObfuscateKey` which will just complicate the transition.