Bugfix: Miner: Don't reuse block_reserved_size for "block is full enough to give up" size delta
Knots 28.1 incorrectly changed a constant of `1000` to `m_options.coinbase_max_additional_size` in the check for when to give up finding another transaction to fill the block:

But this constant did not deal with the reserved size at all. It was in fact simply checking if the block was close to full, and if so, giving up finding another transaction to pad it with after `MAX_CONSECUTIVE_FAILURES` failed attempts.

It doesn't seem very logical to reuse the reserve size for this purpose, and it would be overcomplicated to add yet another setting, so this PR changes it to a new constexpr.

(cherry picked from commit c95d04c8a8e6ba713661bdf0cc5d16d2a30b744a)