bench: measure behavior of a more representative block
https://mempool.space/block/413567 was mined in 2016, added as a benchmark-base in https://github.com/bitcoin/bitcoin/pull/9049.
It lacks modern script types, making the benchmarks unrepresentative of current usage.
In this commit we're replacing it with https://mempool.space/block/784588 from 2023. This block was selected because it's old enough to include legacy script types encountered during IBD, while also containing modern script types in proportions that better reflect current block composition.

bench: document the measured block's properties

This commit documents the current benchmark-base block's properties, to highlight the differences with the replacement block in the next commit.