coins: replace CCoinsCacheEntry::m_flags the m_fresh boolean
A cache entry is now always DIRTY if it is in the linked list.
We can use linked list membership instead of a flag to determine
dirtiness.
Thus, the only state we need to track is if an entry is FRESH.
This can be represented by a simple bool instead of a bitfield.