refactor: inline `CCoinsViewBacked` implementation to header
It's a simple delegate, the cpp is already really crowded, let's do these inline.

Also note that the `CCoinsViewBacked` constructor is asserted to not take any `nullptr` values anymore, since the reason why `base` was stored as a pointer was that references cannot be reallocated, and we need to call `SetBackend` in a few places - which takes a reference, also proving that `nullptr` was never the goal.